remove any types
